PyTorch/generative_models/stable-diffusion-v-2-1/requirements.txt
pytorch-lightning==1.7.7
omegaconf==2.1.1
einops==0.3.0
transformers==4.19.2
open-clip-torch==2.7.0
gradio==3.13.1
k_diffusion==0.0.14
-e .
